refactor(mybatis): 重构 MyBatis 相关代码
- 修改实体类命名规范，将 BO 改为 DO
- 优化分页查询方法，增加排序功能
- 新增忽略逻辑删除的查询和删除方法
- 重构 SQL 注入器，支持自定义 SQL
- 优化 Lambda 查询条件封装，增加空值判断
- 统一数据类型，将 int 改为 Long
- 新增 WebUtils 工具类方法

+++ sofast-spring-boot-starter/sofast-spring-boot-starter-common/src/main/java/cc/sofast/framework/starter/common/dto/PageResult.java
@@ -16,11 +16,11 @@ import java.util.List;
 @NoArgsConstructor
 public class PageResult<T> {
 
-    private int totalCount = 0;
+    private Long totalCount = 0L;
 
-    private int pageSize = 1;
+    private Long pageSize = 1L;
 
-    private int pageIndex = 1;
+    private Long pageIndex = 1L;
 
     private List<T> records;
